Added separate back sink for kafka back up of deleted data
