+++ deleting-compaction-strategy/src/main/java/com/protectwise/cassandra/db/compaction/BackupSinkForDeletingCompaction.java
@@ -15,37 +15,16 @@
  */
 package com.protectwise.cassandra.db.compaction;
 
-import com.protectwise.cassandra.retrospect.deletion.CassandraPurgedData;
-import com.protectwise.cassandra.util.PrintHelper;
-import com.protectwise.cassandra.util.SerializerMetaDataFactory;
-import org.apache.cassandra.config.CFMetaData;
-import org.apache.cassandra.config.ColumnDefinition;
 import org.apache.cassandra.db.*;
 import org.apache.cassandra.db.columniterator.OnDiskAtomIterator;
-import org.apache.cassandra.db.composites.CellNameType;
-import org.apache.cassandra.db.marshal.AbstractType;
-import org.apache.cassandra.db.marshal.CompositeType;
 import org.apache.cassandra.io.sstable.SSTableWriter;
 import org.apache.cassandra.io.sstable.metadata.MetadataCollector;
-import org.apache.cassandra.serializers.SetSerializer;
 import org.apache.cassandra.service.ActiveRepairService;
-import org.apache.cassandra.utils.ByteBufferUtil;
-import org.apache.commons.math3.filter.KalmanFilter;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.codehaus.jackson.map.ObjectMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.nio.charset.CharacterCodingException;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Properties;
-import java.util.function.Consumer;
 
 public class BackupSinkForDeletingCompaction implements IDeletedRecordsSink
 {
@@ -61,16 +40,13 @@ public class BackupSinkForDeletingCompaction implements IDeletedRecordsSink
 	protected long numCells = 0;
 	protected long numKeys = 0;
 
-	private KafkaProducer<String, String> backupRowproducer;
-	private String cassandraPurgedKafkaTopic;
-
-	public BackupSinkForDeletingCompaction(ColumnFamilyStore cfs, File targetDirectory, String kafkaServers, String cassandraPurgedKafkaTopic)
+	public BackupSinkForDeletingCompaction(ColumnFamilyStore cfs, File targetDirectory)
 	{
 		// TODO: Wow, this key estimate is probably grossly over-estimated...  Not sure how to get a better one here.
-		this(cfs, targetDirectory, cfs.estimateKeys() / cfs.getLiveSSTableCount(), kafkaServers, cassandraPurgedKafkaTopic);
+		this(cfs, targetDirectory, cfs.estimateKeys() / cfs.getLiveSSTableCount());
 	}
 
-	public BackupSinkForDeletingCompaction(ColumnFamilyStore cfs, File targetDirectory, long keyEstimate, String kafkaServers, String cassandraPurgedKafkaTopic)
+	public BackupSinkForDeletingCompaction(ColumnFamilyStore cfs, File targetDirectory, long keyEstimate)
 	{
 		this.cfs = cfs;
 		this.targetDirectory = targetDirectory;
@@ -80,135 +56,15 @@ public class BackupSinkForDeletingCompaction implements IDeletedRecordsSink
 		// we don't need to bother resorting the data.
 		columnFamily = ArrayBackedSortedColumns.factory.create(cfs.keyspace.getName(), cfs.getColumnFamilyName());
 
-		this.cassandraPurgedKafkaTopic = cassandraPurgedKafkaTopic;
-
-		Properties kafkaProperties = new Properties();
-		kafkaProperties.setProperty("bootstrap.servers", kafkaServers);
-		kafkaProperties.setProperty("key.serializer",  "org.apache.kafka.common.serialization.StringSerializer");
-		kafkaProperties.setProperty("value.serializer", "org.apache.kafka.common.serialization.StringSerializer");
-		kafkaProperties.setProperty("acks", "1");
-		kafkaProperties.setProperty("retries", "5");
-		kafkaProperties.setProperty("batch.size", "20");
-
-		this.backupRowproducer = new KafkaProducer<String, String>(kafkaProperties);
 	}
 
 	protected void flush()
 	{
 		if (!columnFamily.isEmpty())
 		{
-			//TODO, instead of printRow give a meaningfull name as it will put the purge data in kafka
-			nonLocalRowArchive(this::printCellConsumer);
 			writer.append(currentKey, columnFamily);
 			columnFamily.clear();
 		}
-	}
-
-	private void printCellConsumer(Cell cell) {
-		if (cell != null)  {
-			Cell column = cell;
-			ColumnDefinition columnDefinition = columnFamily.metadata().getColumnDefinition(column.name());
-			if(columnDefinition == null) {
-				return;
-			}
-			if(columnDefinition.isPrimaryKeyColumn()) {
-				logger.info("primary key: {}", column.name().cql3ColumnName(columnFamily.metadata()).toString());
-				return;
-			}
-			try {
-				if (column.value() != null && column.value().array().length > 0) {
-					logger.debug("column identifier:{}", column.name().cql3ColumnName(columnFamily.metadata()).toString());
-					//logger.info("Column value: {}", columnFamily.metadata().getColumnDefinition(column.name()).type.getSerializer().deserialize(column.value()));
-					//logger.info("Column serilizer: {}", columnFamily.metadata().getColumnDefinition(column.name()).type.getSerializer().getClass().getName());
-				}
-			} catch (Exception e) {
-				logger.warn("Exception occurred while printing cell", e);
-			}
-		}
-	}
-
-	private void handleNonLocalArchiving(Cell cell, CassandraPurgedData purgedData) {
-		if (cell != null)  {
-			Cell column = cell;
-			ColumnDefinition columnDefinition = columnFamily.metadata().getColumnDefinition(column.name());
-			if(columnDefinition == null) {
-				return;
-			}
-			if(columnDefinition.isPrimaryKeyColumn()) {
-				logger.debug("primary key: {}", column.name().cql3ColumnName(columnFamily.metadata()).toString());
-				purgedData.addPartitonKey(column.name().cql3ColumnName(columnFamily.metadata()).toString(),
-						SerializerMetaDataFactory.getSerializerMetaData(columnFamily.metadata().getColumnDefinition(column.name()).type.getSerializer()),
-						column.value(), Long.valueOf(column.timestamp()));
-				return;
-			} else {
-				purgedData.addNonKeyCell(SerializerMetaDataFactory.getSerializableCellData(cell, columnFamily), column.name().cql3ColumnName(columnFamily.metadata()).toString(), SerializerMetaDataFactory.getSerializerMetaData(columnFamily.metadata().getColumnDefinition(column.name()).type.getSerializer()));
-
-			}
-		}
-	}
-
-	//TODO: move this to a different backup sink class
-	private void nonLocalRowArchive(Consumer<Cell> printCell) {
-		CassandraPurgedData cassandraPurgedData = new CassandraPurgedData();
-		cassandraPurgedData.setKsName(columnFamily.metadata().ksName);
-		cassandraPurgedData.setCfName(columnFamily.metadata().cfName);
-
-		//retrieve partition key, clustering key value and put in the serializerMetaData
-		handlePartitionKey(currentKey, columnFamily.metadata(), cassandraPurgedData);
-		columnFamily.forEach(cell-> {
-			//logger.info("Clustering keys: {}", PrintHelper.printClusteringKeys(cell, columnFamily.metadata()));
-			handleClusteringKey(cell, columnFamily.metadata(), cassandraPurgedData);
-			handleNonLocalArchiving(cell, cassandraPurgedData);
-			//printCell.accept(cell);
-		});
-
-
-		ObjectMapper objectMapper = new ObjectMapper();
-
-
-		String key = getKafkaMessageKey(columnFamily);
-
-		try {
-			backupRowproducer.send(new ProducerRecord<String, String>(cassandraPurgedKafkaTopic, key, objectMapper.writeValueAsString(cassandraPurgedData)));
-		} catch (Exception e) {
-			logger.warn("Exception occurred while queuing data", e);
-		}
-	}
-
-	private void handleClusteringKey(Cell cell, CFMetaData metadata, CassandraPurgedData cassandraPurgedData) {
-		for (ColumnDefinition def : metadata.clusteringColumns()) {
-			try {
-				cassandraPurgedData.addClusteringKey(ByteBufferUtil.string(def.name.bytes),
-						SerializerMetaDataFactory.getSerializerMetaData(def.type.getSerializer()), cell.name().get(def.position()), null);
-			}catch(Exception e) {
-				throw new RuntimeException(e);
-			}
-		}
-	}
-
-	private void handlePartitionKey(DecoratedKey currentKey, CFMetaData metadata, CassandraPurgedData cassandraPurgedData) {
-		ByteBuffer[] keyParts;
-		AbstractType<?> validator = metadata.getKeyValidator();
-		if (validator instanceof CompositeType) {
-			keyParts = ((CompositeType) validator).split(currentKey.getKey());
-		} else {
-			keyParts = new ByteBuffer[]{
-					currentKey.getKey()
-			};
-		}
-		List<ColumnDefinition> pkc = metadata.partitionKeyColumns();
-		for (ColumnDefinition def : pkc) {
-			try {
-				cassandraPurgedData.addPartitonKey(ByteBufferUtil.string(def.name.bytes),
-						SerializerMetaDataFactory.getSerializerMetaData(def.type.getSerializer()), keyParts[def.position()], null);
-			}catch(Exception e) {
-				throw new RuntimeException(e);
-			}
-		}
-	}
-
-	private String getKafkaMessageKey(ColumnFamily columnFamily) {
-		return String.join("::", columnFamily.metadata().ksName , columnFamily.metadata().cfName);
 	}
 
 	@Override
@@ -266,12 +122,6 @@ public class BackupSinkForDeletingCompaction implements IDeletedRecordsSink
 		{
 			// If deletion convicted nothing, then don't bother writing an empty backup file.
 			abort();
-		}
-
-		try {
-			backupRowproducer.close();
-		}catch(Exception e) {
-			logger.error("Couldn't cleanly stop kafka producer exception ", e);
 		}
 	}
